requirements.txt
pytest
selenium
PyYAML
